Correctly handle hyphenated search terms on Postgres
Fixes #56

The Postgres `to_tsquery` function (used internally by `django.contrib.postgres.search` for SearchQuery) has a quirk/bug when passed a term such as 'poseidon-1234' - it treats this as two tokens 'poseidon' and '-1234', whereas 'poseidon-1234a' gives 'poseidon' and '1234a' which is consistent with how this text is indexed. To bypass this behaviour, treat `-` as a separator when handling PlainText query terms.